Register page and login page switch
